+++ NEWS.md
@@ -5,6 +5,7 @@ Version 1.0.1.9000 (dev)
 ------------------------
 
  - Return summary data in function `boxplot` for objects of class `james`.
+ - Allow to specify number of significant digits in `summary`.
 
 Version 1.0.1 (18/06/2015)
 --------------------------